--- homeyess/templates/map.html
@@ -1,67 +1,0 @@
-<html>
-<body>
-                  <div class="col-sm-offset-1" id="map_canvas" style="height:100%; width:100%;"></div>
-                  <br>    
-<script>
-function loadScript() {
-  var script = document.createElement('script');
-  script.type = 'text/javascript';
-  script.src = "https://maps.googleapis.com/maps/api/js?key={{GOOGLE_MAPS_API_KEY}}&callback=initialize"; 
-  document.body.appendChild(script);
-}
-
-window.onload = loadScript;
-</script>    
-<style>
-  body{
-    text-align:center;
-  }
-  .gmap3{
-    margin: 20px auto;
-    border: 1px dashed #C0C0C0;
-    width: 500px;
-    height: 500px;
-  }
-</style>    
-<script type="text/javascript">
-
-  var map;
-  function initialize() {
-    var mapDiv = document.getElementById('map_canvas');
-    map = new google.maps.Map(mapDiv, {
-      center: new google.maps.LatLng(48.208174,16.373819),
-      zoom: 12,
-      mapTypeId: google.maps.MapTypeId.ROADMAP
-    });
-
-    google.maps.event.addListenerOnce(map, 'tilesloaded', addMarkers);
-
-  }
-  function addMarkers() {
-      var point = new google.maps.LatLng(48.208174,16.373819);
-      var marker = new google.maps.Marker({
-            position: point,
-            map: map,
-            url: 'http://172.16.0.101:8882/zone/1',
-           title: 'test_title',
-        });
-             marker['infowindow']  = new google.maps.InfoWindow({
-                     content: "testing content",
-        });
-            google.maps.event.addListener(marker, 'click', function() {
-                //window.location.href = this.url;
-                 this['infowindow'].open(map, this);
-            });
-           google.maps.event.addListener(marker, 'mouseover', function() {
-                // this['infowindow'].open(map, this);
-                    });
-           google.maps.event.addListener(marker, 'mouseout', function() {
-                // this['infowindow'].close(map, this);
-
-            });
-}
-</script>
-
-</body>
-</html>
-
